LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-item-min-auto-size-001-expected.txt


PASS Computed min-width/min-height of specified auto for grid item. 
FAIL Computed min-width/min-height of specified auto inside display:none which would otherwise have been a grid item. assert_equals: expected "0px" but got "auto"
FAIL Computed min-width/min-height of specified auto for grid item inside display:contents. assert_equals: expected "auto" but got "0px"
FAIL Computed min-width/min-height of specified auto with display:none which would otherwise have been a grid item. assert_equals: expected "0px" but got "auto"
